### 2022-11-09 #4921 @editorialbot
[PRE REVIEW]: Speakerbox: Few-Shot Learning for Speaker Identification with Transformers
**Submitting author:** @evamaxfield (<a href="http://orcid.org/0000-0003-2564-0373">Jackson Brown</a>)
**Repository:** https://github.com/CouncilDataProject/speakerbox
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
