### 2022-11-10 #4921 @evamaxfield
Hey @arfon! I will push a fix for those DOIs.

I also have two potential recommendations for editor:
* https://github.com/pdwaggoner
* https://github.com/cosimameyer

Lastly, can I ask how my deadname was picked in ORCID?

Thanks for all you do!
