+++ paper/paper.bib
@@ -124,7 +124,8 @@
   number={1},
   pages={23--31},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+	doi={10.1002/pra2.601}
 }
 
 @misc{cookiecutter,
@@ -139,8 +140,8 @@
 
 @ARTICLE{SpeakerDia2012,
   author={Anguera, Xavier and Bozonnet, Simon and Evans, Nicholas and Fredouille, Corinne and Friedland, Gerald and Vinyals, Oriol},
-  journal={IEEE Transactions on Audio, Speech, and Language Processing}, 
-  title={Speaker Diarization: A Review of Recent Research}, 
+  journal={IEEE Transactions on Audio, Speech, and Language Processing},
+  title={Speaker Diarization: A Review of Recent Research},
   year={2012},
   volume={20},
   number={2},
@@ -160,6 +161,7 @@
   Author = {{Bredin}, Herv{\'e} and {Laurent}, Antoine},
   Booktitle = {Proc. Interspeech 2021},
   Year = {2021},
+	DOI = {10.21437/interspeech.2021-560},
 }
 
 @inproceedings{Gecko2019,
@@ -177,6 +179,8 @@
   author={Yang, Shu-wen and Chi, Po-Han and Chuang, Yung-Sung and Lai, Cheng-I Jeff and Lakhotia, Kushal and Lin, Yist Y and Liu, Andy T and Shi, Jiatong and Chang, Xuankai and Lin, Guan-Ting and others},
   journal={arXiv preprint arXiv:2105.01051},
   year={2021}
+	doi={10.21437/interspeech.2021-1775},
+
 }
 
 @inproceedings{wolf-etal-2020-transformers,
@@ -196,6 +200,7 @@
 	title        = "VoxCeleb: a large-scale speaker identification dataset",
 	booktitle    = "INTERSPEECH",
 	year         = "2017",
+	doi 				 = "10.21437/interspeech.2017-950",
 }
 
 @article{Wolters2020ASO,
@@ -208,37 +213,37 @@
 
 @misc{li2022-few-shot,
   doi = {10.48550/ARXIV.2204.11180},
-  
+
   url = {https://arxiv.org/abs/2204.11180},
-  
+
   author = {Li, Yanxiong and Wang, Wucheng and Chen, Hao and Cao, Wenchang and Li, Wei and He, Qianhua},
-  
+
   keywords = {Audio and Speech Processing (eess.AS), Sound (cs.SD), FOS: Electrical engineering, electronic engineering, information engineering, FOS: Electrical engineering, electronic engineering, information engineering, FOS: Computer and information sciences, FOS: Computer and information sciences},
-  
+
   title = {Few-Shot Speaker Identification Using Depthwise Separable Convolutional Network with Channel Attention},
-  
+
   publisher = {arXiv},
-  
+
   year = {2022},
-  
+
   copyright = {Creative Commons Attribution 4.0 International}
 }
 
 @misc{kumar2020-few-shot,
   doi = {10.48550/ARXIV.2012.07252},
-  
+
   url = {https://arxiv.org/abs/2012.07252},
-  
+
   author = {Kumar, Neeraj and Goel, Srishti and Narang, Ankur and Lall, Brejesh},
-  
+
   keywords = {Audio and Speech Processing (eess.AS), Machine Learning (cs.LG), FOS: Electrical engineering, electronic engineering, information engineering, FOS: Electrical engineering, electronic engineering, information engineering, FOS: Computer and information sciences, FOS: Computer and information sciences},
-  
+
   title = {Few Shot Adaptive Normalization Driven Multi-Speaker Speech Synthesis},
-  
+
   publisher = {arXiv},
-  
+
   year = {2020},
-  
+
[…]
