### 2022-11-14 6539d1b
close command in CiA citation
