+++ paper/paper.bib
@@ -125,7 +125,7 @@
   pages={23--31},
   year={2022},
   publisher={Wiley Online Library},
-	doi={10.1002/pra2.601}
+  doi={10.1002/pra2.601},
 }
 
 @misc{cookiecutter,
