### 2022-11-14 ccfd145
Fix DOIs flagged by Whedon  (#11)
* push missing dois

* close command in CiA citation
