+++ paper/paper.bib
@@ -178,9 +178,8 @@
   title={SUPERB: Speech processing Universal PERformance Benchmark},
   author={Yang, Shu-wen and Chi, Po-Han and Chuang, Yung-Sung and Lai, Cheng-I Jeff and Lakhotia, Kushal and Lin, Yist Y and Liu, Andy T and Shi, Jiatong and Chang, Xuankai and Lin, Guan-Ting and others},
   journal={arXiv preprint arXiv:2105.01051},
-  year={2021}
-	doi={10.21437/interspeech.2021-1775},
-
+  year={2021},
+  doi={10.21437/interspeech.2021-1775},
 }
 
 @inproceedings{wolf-etal-2020-transformers,
