### 2022-11-14 #4921 @evamaxfield
DOIs should now be fixed!
