### 2022-11-15 #4921 @gkthiruvathukal
Hi, this is the new track editor, @gkthiruvathukal checking in. I am working to get this and a growing number of track submissions reviewed. Please do bear with me (and JOSS) as there is a bit of a backlog with many of our editors at maximum number of assignments they can edit at the moment.  Thank you for your patience and understanding.
