+++ speakerbox/preprocess.py
@@ -21,6 +21,7 @@ from .utils import set_global_seed
 
 log = logging.getLogger(__name__)
 
+
 ###############################################################################
 # Annotation
 
@@ -403,7 +404,10 @@ def expand_labeled_diarized_audio_dir_to_dataset(
         # Iter over all labels
         for label_dir in anno_dir.iterdir():
             if label_dir.is_dir():
-                for i, audio_file in enumerate(label_dir.glob("*.wav")):
+                glob = label_dir.glob
+                for i, audio_file in enumerate(
+                    glob("*.wav") + glob("*.mp3") + glob("*.aac") + glob("*.ogg") + glob("*.raw")
+                    + glob("*.pcm") + glob("*.mp4") + glob("*.flv")):
                     audio = AudioSegment.from_file(audio_file)
                     monologue_speaker = label_dir.name
 
