### 2023-01-11 #4921 @faroit
👋  Hi @evamaxfield, thanks again for your submission to JOSS! I will finally be starting the review process now. Sorry again for the delay that I have caused on my side. I am usually trying to aim for a quick review period so this isn't something I am happy with. I will try my best now to speed things up.

I had a look at your reviewer suggestions - thanks. I will ping them in the follow up comment. In addition I would also like to see a review from someone with domain expertise in audio processing. I would suggest

@ jonashaag
@ adefossez
@ jordipons
@ santi-pdp
@ TParcollet
@ popcornell

For that list, please check the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and tell me if I can ping them here.
