### 2023-01-17 #4921 @faroit
👋  @jonashaag @adefossez @jordipons @santi-pdp @TParcollet @popcornell @boeddeker - would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is [speakerbox](https://github.com/CouncilDataProject/speakerbox).
