### 2023-01-29 #4921 @nniiicc
No problem! Here are some recent JOSS authors that likely have expertise:
- `@pbotros`
- `@Ayuei`
- '@FabianHofmann'
- `@desilinguist`
