### 2023-01-31 #4921 @FabianHofmann
@faroit I think from a content perspective the paper does not fit to me, as I am not experienced with recording and sound tuning (but I could serve as a backup if no one is found)
