### 2023-02-07 #5132 @editorialbot
[REVIEW]: Speakerbox: Few-Shot Learning for Speaker Identification with Transformers
**Submitting author:** @evamaxfield (<a href="http://orcid.org/0000-0003-2564-0373">Eva Maxfield Brown</a>)
**Repository:** https://github.com/CouncilDataProject/speakerbox
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
