### 2023-03-02 #5132 @desilinguist
This is a well-written, well-engineered submission that is likely to be useful to the larger NLP community. Speaker identification and diarization are common problems encountered in many applications and an easy-to-use Python package is a great contribution.

I only have a few minor suggestions:

1. Provide a [conda](https://conda.io) package for installation as conda is a pretty popular package management option in the NLP/ML field.
2. Provide a tutorial so that someone who's not very familiar with the field or the tool can easily try it out without using a GPU or spending a lot of time. I couldn't really test the package beyond installation since there was no such walkthrough/tutorial.
3. Can the underlying model take advantage of an Apple Silicon GPU? Given the popularity of those machines, it may be worth adding another table if it can. If not, a brief note indicating which GPUs are "officially" supported would be useful.
