### 2023-03-02 #5132 @FabianHofmann
I found some time to read the paper and I agree, it is well-written and understandable even for non-experts like me:

Some comments about the paper:

- Is the term "coporara" well-known in the field (or is it a typo and should be "corpora")? I could not find any definition after a very short research.
- As a non-expert, I asked myself what "Wav2Vec2 Transformer" should mean
- In line 28, it should be "its" not "it's"
- I did not fully understand the sentence from line 29 - 31, in particular what does "for evidence" mean here?
- In line 61, it should be "their"
- The caption in the figure sounds a bit off "Workflow of using various functions provided by Speakerbox to prepare a dataset and fine-tune a new speaker identification model". Perhaps "Scheme of typical workflow to prepare a dataset and fine-tune a new speaker identification model with Speakerbox"
- In general, I feel the font in the image is a bit small. Could you perhaps go for a vertical layout of the boxes to increase the font size? Also, if I see correctly, the code in the box does not reference to the `wav` files in the `example-traning-data` directory, which I find a bit confusing.

Otherwise it looks well!
