### 2023-03-03 #5132 @evamaxfield
Hello @desilinguist and @FabianHofmann! Thanks for going through the repo and the paper.

Some answers:

@desilinguist
> Provide a tutorial so that someone who's not very familiar with the field or the tool can easily try it out without using a GPU or spending a lot of time. I couldn't really test the package beyond installation since there was no such walkthrough/tutorial.

I constructed the README as a set of vignettes -- each shows how to use the library and the functionality of different library components -- Starting at [Workflow](https://github.com/CouncilDataProject/speakerbox#workflow) it goes through diarization, cleaning, and training a model. If there is something more you are looking for happy to add it.

We additionally give functionality examples in the [Problem](https://github.com/CouncilDataProject/speakerbox#problem) section of the README, with additional reproducible performance benchmarks following the directions at the end of that section.

> Can the underlying model take advantage of an Apple Silicon GPU? Given the popularity of those machines, it may be worth adding another table if it can. If not, a brief note indicating which GPUs are "officially" supported would be useful.

The broad answer is "any GPU capable of using pytorch / CUDA." I don't personally think I should have a table of all of those because it is a different library doing the work there but would it be okay if I link to the pytorch page somehow?

@FabianHofmann
> Is the term "coporara" well-known in the field (or is it a typo and should be "corpora")? I could not find any definition after a very short research.

Typo! I will fix it!

> As a non-expert, I asked myself what "Wav2Vec2 Transformer" should mean.

Hmmmm. I actually don't think that additional info is adding much. I am tempted to remove it if that is okay?

> I did not fully understand the sentence from line 29 - 31, in particular what does "for evidence" mean here?

We can expand this / rework this sentence. The point we were trying to get across is that in legal court proceedings, journalistic inquiry, etc. it is crucial to have speakers labeled because you can cite the labeled recording as evidence.

> The caption in the figure sounds a bit off "Workflow of using various functions provided by Speakerbox to prepare a dataset and fine-tune a new speaker identification model". Perhaps "Scheme of typical workflow to prepare a dataset and fine-tune a new speaker identification model with Speakerbox"

That makes sense to me! Will update it to something like that.

> In general, I feel the font in the image is a bit small. Could you perhaps go for a vertical layout of the boxes to increase the font size? Also, if I see correctly, the code in the box does not reference to the wav files in the example-traning-data directory, which I find a bit confusing.

If you could clarify what you mean. What "code in the box". The code snippet on the right side of the workflow diagram?
