### 2023-03-03 #5132 @desilinguist
> I constructed the README as a set of vignettes -- each shows how to use the library and the functionality of different library components -- Starting at [Workflow](https://github.com/CouncilDataProject/speakerbox#workflow) it goes through diarization, cleaning, and training a model. If there is something more you are looking for happy to add it.

I see! I guess as someone who doesn't really work in this area, it wasn't clear to me that's what it was. I usually look for a tutorial or quick start that's really easy to jump into and get a result with – but this seemed a bit more involved (e.g., downloading something from Google Drive). May be it's just a me problem but I would have liked a clearly labeled section with commands (without a lot of text between them) that can be run with some sample data already part of the repo. Just my $0.02.

> We additionally give functionality examples in the [Problem](https://github.com/CouncilDataProject/speakerbox#problem) section of the README, with additional reproducible performance benchmarks following the directions at the end of that section.

Yes but this seems to require a GPU and would run in about an hour which is a little too long for a quick walkthrough.

> The broad answer is "any GPU capable of using pytorch / CUDA." I don't personally think I should have a table of all of those because it is a different library doing the work there but would it be okay if I link to the pytorch page somehow?

That's fair enough.
