+++ speakerbox/types.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 from dataclasses import dataclass
 from pathlib import Path
