+++ speakerbox/tests/__init__.py
@@ -1,3 +1,1 @@
-# -*- coding: utf-8 -*-
-
 """Unit test package for speakerbox."""
