+++ speakerbox/tests/test_preprocess.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 from pathlib import Path
 
 import pytest
 
 from speakerbox import preprocess
+
 
 ###############################################################################
 
