+++ speakerbox/tests/data/__init__.py
@@ -1,3 +1,1 @@
-# -*- coding: utf-8 -*-
-
-"""Data for unit tests for speakerbox"""
+"""Data for unit tests for speakerbox."""
