### 2023-03-07 #5132 @evamaxfield
@desilinguist I completely understand where you are coming from with the desire to have a quickstart. I look for that when I am looking at a new repo too.

But I don't know how to construct the README to allow for such a quickstart when this library isn't for tabular data -- its for audio data which typically does take a long time to annotate and train against. That could be my lack of knowledge though.

I have thought about this a decent amount and I think one of the ways which I can have a better example of this library's usage would be to record a three minute video of me going through the whole process. Would that be okay for you?
