### 2023-03-08 c4d2c07
Fix links in README, add a bit more to code snippets
