### 2023-03-08 e80cf24
Log the value counts if you don't have enough data
