### 2023-03-09 #5132 @evamaxfield
@desilinguist It isn't quite a 3 minutes video but I made a video which shoes a complete run-through of the library and added it to the [README](https://github.com/councildataproject/speakerbox#example-usage-video).

You can watch the video here: https://youtu.be/SK2oVqSKPTE

Hopefully that video helps explain why we went with the vignette style README. I do think the video is a nice addition to the documentation though!
