### 2023-03-09 #5132 @FabianHofmann
Hey @evamaxfield I think the paper looks good now!

> If you could clarify what you mean. What "code in the box". The code snippet on the right side of the workflow diagram?

Yes, this is what I meant. It looks not professional if the font size in the diagram is that much smaller. Would a vertical layout of the charts be possible?

The code base looks very good.  Two remaining points to make things more elegant.

1. The README on GH is quite large. Many projects suggest to keep the README clean and move all important things to the documentation. (I'd keep the video as it's very nice)

2. The example
```
from speakerbox.examples import train_and_eval_all_example_models

results = train_and_eval_all_example_models(
    # be sure to provide the correct path to the unzipped directory
    "/home/eva/Downloads/example-speakerbox-dataset/",
)
```
works like charme. I'd replace "/home/eva/Downloads/example-speakerbox-dataset/" by "~/Downloads/example-speakerbox-dataset/". In the long term, it would perhaps make sense to create a `io` module for retrieving and storing (test) data. Like this you would enable the code snippet to  be be runable as is. I have something in mind like this
```
from speakerbox.examples import train_and_eval_all_example_models, get_example_dataset

path = get_example_dataset()
results = train_and_eval_all_example_models(path)
```
The `get_example_dataset()` downloads and unzips in the background using the `io` module. This is a suggestion not a must have :)

Otherwise everything looks good!
