### 2023-03-09 e900037
Add new function to download and unpack the example dataset for reproducible models
