+++ pyproject.toml
@@ -50,6 +50,9 @@ Documentation = "https://CouncilDataProject.github.io/speakerbox"
 # extra dependencies
 # https://peps.python.org/pep-0621/#dependencies-optional-dependencies
 [project.optional-dependencies]
+example_data = [
+  "gdown~=4.6",
+]
 lint = [
   "black>=22.3.0",
   "cdp-backend~=3.0.3",
