### 2023-03-09 #5132 @evamaxfield
Another round of updates!

@FabianHofmann I hope I have resolved two of your three requests:

1. I updated the workflow image to be vertical and with large font size (and with additional clarity imo). https://github.com/CouncilDataProject/speakerbox#problem

2. The code to reproduce the models in the table now includes a function to automatically download the data instead of forcing the reader to go to google drive:

Screenshot:
![Screenshot 2023-03-09 at 11-57-59 Address most of README feedback · CouncilDataProject_speakerbox@2dcc0be](https://user-images.githubusercontent.com/17132317/224140227-484d4998-0021-4ee3-8af2-69cef56b83bc.png)

---

If I can, I want to push back on removing content from the README. From my personal experience, users tend to only search the README rather than the larger projects documentation. I have had a lot of GitHub issues and Slack messages with "how can I do this" when the details were in the larger docs and not in the README when working on other projects.

I hope it is okay if keep the README in is comprehensive format.
