### 2023-03-09 #5132 @evamaxfield
Running through the list of items remaining for each reviewer -- what can I do to help complete the checklist for each of these?

@desilinguist:
* Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
* Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

@FabianHofmann:
* Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
* Functionality: Have the functional claims of the software been confirmed?
* Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
* Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
* References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

Please let me know! Happy to add things as needed.
