### 2023-03-09 #5132 @FabianHofmann
@evamaxfield thank you for being so responsive! I agree with your concerns about changing the README. Please note that the new figure is somehow too large for the figure caption (have a look at the latest article). Otherwise everything is very nice :) Consider the paper also approved from my side, well done!
