### 2023-03-09 #5132 @evamaxfield
Thank you both!

I am going to quickly try and fix the figure sizing issue. Otherwise @faroit seems like we can move this process along to your end!
