### 2023-03-09 2500ee4
Fit image and fig text within a single page
