### 2023-03-09 #5132 @evamaxfield
The figure has been updated to just barely fit on a single page! Let me know what to do next @faroit

Thanks again all :heart:
