+++ paper/paper.md
@@ -134,7 +134,7 @@ The few-shot approach we have described in Speakerbox can be extended, and made 
 
 1. **The creation of a GUI for dataset preparation, training, and application:** GUIs developed with Python have become much more accessible in recent years with tools such as [MagicGUI](https://github.com/napari/magicgui) being developed. A GUI would likely help non-computational scientists more easily train their own fine-tuned Speakerbox models.
 
-2. **The creation of a template repository to assist in the "productization" of Speakerbox models:** The construction of a "template" or "cookiecutter" [@cookiecutter] repository may also be useful. A template repository may provide a standard configuration for the storage and versioning of the training data and the management of model training and evaluation via continuous integration. We plan to create and embed a similar system within Council Data Project (CDP) infrastructures to enable the transparent training of speaker identification models specific to each municipal council. In our case, we plan to have a continuous integration attempt to train a new model when new training data is added to the infrastructure repository. If the trained model reaches an accuracy threshold, we plan to store this model on our remote storage system. This model would then be used to annotate transcripts with their speakers during the processing of each municipal meeting.
+2. **The creation of a template repository to assist in the "productization" of Speakerbox models:** The construction of a "template" or "cookiecutter" [@cookiecutter] repository may also be useful. A template repository may provide a standard configuration for the storage and versioning of the training data and the management of model training and evaluation via continuous integration. We plan to create and embed a similar system within Council Data Project (CDP) infrastructures to enable the transparent training of speaker identification models specific to each municipal council. In our case, we plan to have a continuous integration system attempt to train a new model when new training data is added to the infrastructure repository. If the trained model reaches an accuracy threshold, we plan to store this model on our remote storage system. This model would then be used to annotate transcripts with their speakers during the processing of each municipal meeting.
 
 # Acknowledgements
 
