### 2023-03-13 9812c1d
Ignore zenodo file in check-manifest
