+++ paper/paper.md
@@ -23,7 +23,7 @@ affiliations:
     - name: University of Washington, Seattle
       index: 2
 
-date: 7 September 2022
+date: 13 March 2023
 bibliography: paper.bib
 ---
 
