### 2023-03-13 #5132 @evamaxfield
Hey @faroit!

* New release (all for version v1.2.0):
  * GitHub Release: https://github.com/CouncilDataProject/speakerbox/releases/tag/v1.2.0
  * GitHub Tag: https://github.com/CouncilDataProject/speakerbox/releases/tag/v1.2.0
  * PyPI Release: https://pypi.org/project/speakerbox/1.2.0/
* Archive (auto-released from v1.2.0 CI):
  * Zenodo: https://zenodo.org/record/7729994
  * Zenodo DOI: 10.5281/zenodo.7729994
