+++ CITATION.cff
@@ -82,3 +82,5 @@ preferred-citation:
   issue: 83
   volume: 8
   year: 2023
+  url: "https://doi.org/10.21105/joss.05132"
+  publisher: "The Open Journal"
